requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
decorator==5.1.1
executing==2.1.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gigachat==0.1.36
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
ipython==8.29.0
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-gigachat==0.3.1
langchain-text-splitters==0.3.2
langgraph==0.2.53
langgraph-checkpoint==2.0.6
langgraph-sdk==0.1.36
langsmith==0.1.146
marshmallow==3.23.1
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.12
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pypdf==5.1.0
python-dotenv==1.0.1
python-telegram-bot==21.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
Telethon==1.37.0
tenacity==9.0.0
tokenizers==0.20.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.0
